requirements.txt
Flask==3.1.2
flask_jwt_extended==4.7.1
pandas==2.3.3
python-dotenv==1.1.1
python_dateutil==2.9.0.post0
Werkzeug==3.1.3
gunicorn==23.0.0
openpyxl==3.1.5 
XlsxWriter==3.2.9
